Removed --tree option in profile show
