+++ requirements.txt
@@ -3,6 +3,5 @@ termcolor>=1.1.0
 texttable>=1.6.2
 python-ptrace>=0.9.7
 pyelftools==0.27
-tree-format
 requests==2.26.0
 tinydb==4.5.2
